Project init with GUI class to draw the GUI
GUI class implements tkinter to draw the window with a menu bar. The menu
bar has an action that calls a function from the generate module which will
be filled out later to generate a new maze.

The GUI class has a canvas area that will be used to render the generated
mazes and will display the solution to a given maze.