generate module draws a graph and displays it in the gui window
First, the generate module. This is using a procedural generation algorithm
to generate a maze. The algorithm starts with an empty (zeroed) 2d numpy
array and recursively fills it with a path by way of a backtracking
algorithm. The algorithm draws a path by randomly selecting a direction to
explore the grid from a starting point carving a path. The traversal itself
is in a 3*3 slice so it maintains a wall on either side of te path.

There is no guarantee that the algorithm will generate a valid maze every.
But I suspect this is a bug in the placement of the outermost walls by the
border function. I'll investigate that issue to see if I can get consistently
procedurally generated mazes before moving onto a path traversal algorithm
to calculate a path through the maze.

The gui module renders the chart by casting the maze array to a matplotlib
Figure and then rendering it in a tkinter canvas. The canvas is then packed
into a tkinter frame and the frame is packed into the root window.

The root window is reloaded to refresh the view. If a new maze is generated
the previous maze is destroyed prior to the new maze being mounted in the view.