Tweaked the border function.
Made some small improvements. It will always generate an empty exit point around
the maze border on both sides at the correct point.

Also tweaked the gui maze height and width to by 75*75. This is best for even
rendering of the maze and complex enough without incurring the wrath of the
max recursion depth on my machine.

The maze may not always be solvable. Sometimes the exit is just boxed in.
Events like this should be factored into the solver solution.