Updated the readme.
Just a little paragraph about the project, and a bit on how to run it.

note: at this point, the readme says install the requirements but the requirements.txt file is empty.
so have fun with that.

For reference the current stack you'd need to install are numpy and matplotlib.
that should be it.