Added a maze solver algorithm.
Made a new script in isolation that is to be implemented as a module of the
project.

This script recursively traverses the array creating a tree-like structure
of the possible paths. If the path has off-shoots, these are considered children
of the parent node (current path) and are added to the tree. If the path has
no children there is a calculation to determine if its a dead end.

The end of the maze is determined by the dead-end status. A dead end is where the
end of the path is not the end of the maze.

The algorithm is not perfect and does not always find the shortest path. But
it should completely explore the maze and return as many paths as possible that
reach the end.

This is to be implemented with a callback to the gui module to update the rendered
image with the progress by painting the paths explored and the final path.