Maze solver creates a maze and the gui indicates its solving progress.
The solver should update the progress by generating a new image of the updated
maze.

There is an unknown bug that I cannot track down. For some reason, when presented
with a solvable maze, maze explorer paint paths function loop ends, the image
in the GUI reverts to an older state from before the paint function started
(the entry and exit points are painted).

This is only happening on solvable mazes. If presented with an unsolvable maze,
the rendered image does NOT revert.

Its a curious bug and I can't for the life of me figure out why its happening.